# pkg/api/openapi/api.yml
openapi: 3.0.3
info:
  title: PDF Generator API
  description: API para generar y visualizar PDFs
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Servidor local

paths:
  /generate:
    post:
      summary: Genera un PDF
      description: Recibe datos en JSON y genera un PDF
      operationId: generatePDF
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateRequest"
      responses:
        "200":
          description: PDF generado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "PDF generado correctamente"
        "400":
          description: Error en la solicitud
        "405":
          description: Método no permitido

  /view:
    get:
      summary: Visualiza el PDF generado
      description: Retorna el archivo PDF generado
      operationId: viewPDF
      responses:
        "200":
          description: Retorna el PDF generado
          content:
            application/pdf:
              schema:
                type: string
                format: binary

  /pdf/ventas:
    post:
      summary: Genera un PDF del reporte de ventas
      description: Recibe un listado de ventas en JSON y genera un PDF
      operationId: generateReporteVentas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReporteVentasRequest"
      responses:
        "200":
          description: PDF de ventas generado correctamente
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "400":
          description: Error en la solicitud
        "405":
          description: Método no permitido

  /pdf/inventario:
    post:
      summary: Genera un PDF del reporte de inventario
      description: Recibe un listado de inventario en JSON y genera un PDF
      operationId: generateReporteInventario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReporteInventarioRequest"
      responses:
        "200":
          description: PDF de inventario generado correctamente
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "400":
          description: Error en la solicitud
        "405":
          description: Método no permitido

  /pdf/usuarios-activos:
    post:
      summary: Genera un PDF del reporte de usuarios activos
      description: Recibe un listado de usuarios activos en JSON y genera un PDF
      operationId: generateReporteUsuariosActivos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReporteUsuariosActivosRequest"
      responses:
        "200":
          description: PDF de usuarios activos generado correctamente
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "400":
          description: Error en la solicitud
        "405":
          description: Método no permitido

components:
  schemas:
    TemplateRequest:
      type: object
      properties:
        data:
          type: object
          properties:
            nombre:
              type: string
            cargo:
              type: string
            ubicacion:
              type: string
            obra:
              type: string
            tramite:
              type: string
            name:
              type: string
            num_folio:
              type: string
            num_acta:
              type: string
            levantada:
              type: string
            dd:
              type: string
            mmmm:
              type: string
            yyyy:
              type: string
            mm:
              type: string
            url:
              type: string
              format: uri
            template:
              type: string

    Venta:
      type: object
      properties:
        fecha:
          type: string
          example: "2025-09-29"
        producto:
          type: string
          example: "Producto A"
        cantidad:
          type: integer
          example: 2
        precio:
          type: number
          format: float
          example: 150.50

    ReporteVentasRequest:
      type: object
      properties:
        ventas:
          type: array
          items:
            $ref: "#/components/schemas/Venta"

    InventarioItem:
      type: object
      properties:
        id:
          type: string
          example: "INV-001"
        producto:
          type: string
          example: "Laptop"
        categoria:
          type: string
          example: "Electrónica"
        cantidad:
          type: integer
          example: 10
        precio:
          type: number
          format: float
          example: 12000.00

    ReporteInventarioRequest:
      type: object
      properties:
        inventario:
          type: array
          items:
            $ref: "#/components/schemas/InventarioItem"

    UsuarioActivo:
      type: object
      properties:
        id:
          type: string
          example: "USR-123"
        nombre:
          type: string
          example: "Juan Pérez"
        email:
          type: string
          example: "juan.perez@example.com"
        activoDesde:
          type: string
          format: date
          example: "2025-09-01"
        activo:
          type: boolean
          example: true

    ReporteUsuariosActivosRequest:
      type: object
      properties:
        usuarios:
          type: array
          items:
            $ref: "#/components/schemas/UsuarioActivo"